Add RequestPossibleBloodTypes table and Request entity new fields

+++ UnaPinta.Data/Contracts/IUnaPintaRepository.cs
@@ -15,12 +15,12 @@ namespace UnaPinta.Data.Contracts
         void UpdateUser(User user);
         Task<IEnumerable<User>> GetAllDonors();
         Task<IEnumerable<User>> GetDonorsByBloodType(List<BloodTypeEnum> bloodTypes);
-        Task<User> GetUserById(int id);
+        Task<User> GetUserById(long id);
         Task<bool> GetUserByEmail(string email);
 
         //Request table methods
         void CreateRequest(Request request);
-        Task<Request> GetRequestById(int id);
+        Task<Request> GetRequestById(long id);
 
         //BloodTypes table methods
         Task<IEnumerable<BloodType>> GetAllBloodTypes();
@@ -32,11 +32,11 @@ namespace UnaPinta.Data.Contracts
 
         //ConfirmationCodes table methods
         void AddConfirmationCode(ConfirmationCode code);
-        Task<ConfirmationCode> GetCodeByUser(string code, int id);
+        Task<ConfirmationCode> GetCodeByUser(string code, long id);
 
         //WaitList table methods
         void AddWaitListItem(WaitList item);
-        Task<DateTime> GetAvailabilityDateByDonorId(int id);
+        Task<DateTime> GetAvailabilityDateByDonorId(long id);
 
         //Condition table methods
         Task<Condition> GetConditionById(ConditionEnum id);
